pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "llm-eval-env"
version = "1.0.0"
description = "LLM Evaluation Pipeline Environment for OpenEnv"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]
